backend/src/scrapers/database_scraper/requirements.txt
boto3==1.41.1
curl_cffi==0.13.0
itemadapter==0.12.2
psycopg2==2.9.11
python-dotenv==1.2.1
scrapy==2.13.4
<<<<<<< HEAD
scrapy_playwright
=======
scrapy-playwright
>>>>>>> main
